requirements.txt
torch
transformers
Pillow
opencv-contrib-python-headless
pandas
tqdm
albumentations
scikit-learn